+++ src/main/java/com/brandonlenz/iso8583/messages/Iso8583Message.java
@@ -1,6 +1,6 @@
 package com.brandonlenz.iso8583.messages;
 
-import com.brandonlenz.iso8583.definitions.fields.BitmapDefinition;
+import com.brandonlenz.generic.messages.Message;
 import com.brandonlenz.iso8583.definitions.messages.Iso8583MessageDefinition;
 import com.brandonlenz.iso8583.fields.Bitmap;
 import com.brandonlenz.iso8583.fields.DataField;
@@ -14,15 +14,29 @@ import java.util.TreeMap;
 public class Iso8583Message implements Message {
 
     private final Iso8583MessageDefinition definition;
-    private final SortedMap<Integer, DataField> dataFields;
     private DataField messageTypeIndicator;
     private Bitmap primaryBitmap;
+    private final SortedMap<Integer, DataField> dataFields;
 
-    private Iso8583Message(Builder builder) {
-        this.definition = builder.definition;
-        this.messageTypeIndicator = builder.messageTypeIndicator;
-        this.primaryBitmap = builder.primaryBitmap;
-        this.dataFields = builder.dataFields;
+    Iso8583Message(Iso8583MessageDefinition definition) {
+        this.definition = definition;
+        this.dataFields = new TreeMap<>();
+    }
+
+    void setMessageTypeIndicator(DataField messageTypeIndicator) {
+        this.messageTypeIndicator = messageTypeIndicator;
+    }
+
+    void setPrimaryBitmap(Bitmap primaryBitmap) {
+        this.primaryBitmap = primaryBitmap;
+    }
+
+    void setField(int fieldNumber, DataField dataField) {
+        dataFields.put(fieldNumber, dataField);
+    }
+
+    void removeField(int fieldNumber) {
+        dataFields.remove(fieldNumber);
     }
 
     @Override
@@ -87,11 +101,11 @@ public class Iso8583Message implements Message {
         }
     }
 
-    private Optional<Bitmap> getSecondaryBitmap() {
+    Optional<Bitmap> getSecondaryBitmap() {
         return getBitmapField(definition.getSecondaryBitmapFieldNumber());
     }
 
-    private Optional<Bitmap> getTertiaryBitmap() {
+    Optional<Bitmap> getTertiaryBitmap() {
         return getBitmapField(definition.getTertiaryBitmapFieldNumber());
     }
 
@@ -122,131 +136,11 @@ public class Iso8583Message implements Message {
         return getData();
     }
 
-    public static class Builder {
+    public static Iso8583MessageBuilder builder(Iso8583MessageDefinition definition, DataField messageTypeIndicator) {
+        return new Iso8583MessageBuilder(definition, messageTypeIndicator);
+    }
 
-        private final Iso8583MessageDefinition definition;
-        private final DataField messageTypeIndicator;
-        private final Bitmap primaryBitmap;
-        private final SortedMap<Integer, DataField> dataFields = new TreeMap<>();
-
-        public Builder(Iso8583MessageDefinition iso8583MessageDefinition, DataField messageTypeIndicator) {
-            this.definition = iso8583MessageDefinition;
-            this.messageTypeIndicator = messageTypeIndicator;
-            this.primaryBitmap = initializeEmptyBitmap(definition.getPrimaryBitmapDefinition());
-        }
-
-        public Builder setField(int fieldNumber, byte[] rawData) {
-            DataField dataField = definition.getFieldDefinition(fieldNumber)
-                    .getDataFieldBuilder()
-                    .build(rawData);
-            return setField(fieldNumber, dataField);
-        }
-
-        public Builder setField(int fieldNumber, String data) {
-            DataField dataField = definition.getFieldDefinition(fieldNumber)
-                    .getDataFieldBuilder()
-                    .build(data);
-            return setField(fieldNumber, dataField);
-        }
-
-        public Builder setField(int fieldNumber, DataField dataField) {
-            //TODO: Do not allow setting of bitmap fields, as these are automatic
-            if (!definition.getFieldDefinitions().containsKey(fieldNumber)) {
-                throw new IllegalArgumentException(
-                        "DataField number " + fieldNumber + " is not defined for this Message");
-            }
-
-            setCorrespondingBitmapBit(fieldNumber);
-            dataFields.put(fieldNumber, dataField);
-
-            return this;
-        }
-
-        private void setCorrespondingBitmapBit(int fieldNumber) {
-            if (definition.primaryBitmapGovernsBit(fieldNumber)) {
-                primaryBitmap.setBit(fieldNumber);
-            } else if (definition.secondaryBitmapGovernsBit(fieldNumber)) {
-                setSecondaryBitmapBit(fieldNumber);
-            } else if (definition.tertiaryBitmapGovernsBit(fieldNumber)) {
-                setTertiaryBitmapBit(fieldNumber);
-            } else {
-                throw new IllegalArgumentException(
-                        "DataField number " + fieldNumber + " is not governed by any defined bitmap");
-            }
-        }
-
-        private void setSecondaryBitmapBit(int fieldNumber) {
-            setBitmapFieldBit(definition.getSecondaryBitmapFieldNumber(), fieldNumber);
-        }
-
-        private void setTertiaryBitmapBit(int fieldNumber) {
-            setBitmapFieldBit(definition.getSecondaryBitmapFieldNumber(), fieldNumber);
-        }
-
-        private void setBitmapFieldBit(int bitmapFieldNumber, int fieldNumber) {
-            if (dataFields.containsKey(bitmapFieldNumber)) {
-                Bitmap bitmap = (Bitmap) dataFields.get(bitmapFieldNumber);
-                bitmap.setBit(fieldNumber);
-            } else {
-                Bitmap bitmap = initializeEmptyBitmapField(bitmapFieldNumber);
-                bitmap.setBit(fieldNumber);
-                setField(bitmapFieldNumber, bitmap);
-            }
-        }
-
-        private Bitmap initializeEmptyBitmapField(int bitMapFieldNumber) {
-            BitmapDefinition bitmapDefinition = (BitmapDefinition) definition
-                    .getFieldDefinition(bitMapFieldNumber);
-            return initializeEmptyBitmap(bitmapDefinition);
-        }
-
-        private Bitmap initializeEmptyBitmap(BitmapDefinition bitmapDefinition) {
-            return bitmapDefinition.getDataFieldBuilder().build(new byte[bitmapDefinition.getByteLength()]);
-        }
-
-        public Builder removeField(int fieldNumber) {
-            //TODO: Do not allow removal of bitmap fields, as these are automatic
-            if (!dataFields.containsKey(fieldNumber)) {
-                throw new IllegalArgumentException("Message does not contain DataField number " + fieldNumber);
-            }
-
-            unsetCorrespondingBitmapBit(fieldNumber);
-            dataFields.remove(fieldNumber);
-
-            return this;
-        }
-
-        private void unsetCorrespondingBitmapBit(int fieldNumber) {
-            if (definition.primaryBitmapGovernsBit(fieldNumber)) {
-                primaryBitmap.unsetBit(fieldNumber);
-            } else if (definition.secondaryBitmapGovernsBit(fieldNumber)) {
-                unsetSecondaryBitmapBit(fieldNumber);
-            } else if (definition.tertiaryBitmapGovernsBit(fieldNumber)) {
-                unsetTertiaryBitmapBit(fieldNumber);
-            } else {
-                throw new IllegalArgumentException(
-                        "DataField number " + fieldNumber + " is not governed by any defined bitmap");
-            }
-        }
-
-        private void unsetSecondaryBitmapBit(int fieldNumber) {
-            unsetBitmapFieldBit(definition.getSecondaryBitmapFieldNumber(), fieldNumber);
-        }
-
-        private void unsetTertiaryBitmapBit(int fieldNumber) {
-            unsetBitmapFieldBit(definition.getTertiaryBitmapFieldNumber(), fieldNumber);
-        }
-
-        private void unsetBitmapFieldBit(int bitmapFieldNumber, int fieldNumber) {
-            Bitmap bitmap = (Bitmap) dataFields.get(bitmapFieldNumber);
-            bitmap.unsetBit(fieldNumber);
-            if (bitmap.isEmpty()) {
-                removeField(bitmapFieldNumber);
-            }
-        }
-
-        public Iso8583Message build() {
-            return new Iso8583Message(this);
-        }
+    public static Iso8583MessageParser parser(Iso8583MessageDefinition definition) {
+        return new Iso8583MessageParser(definition);
     }
 }
